/profile cmd and /addGold cmd

+++ src/main/resources/schema.sql
@@ -12,7 +12,7 @@ CREATE TABLE IF NOT EXISTS users (
     id INTEGER,
     telegram_id BIGINT,
     username VARCHAR,
-    gold_balance DOUBLE,
+    gold_balance INTEGER,
     real_usd_balance DOUBLE,
     expected_usd_total DOUBLE,
     authorized BOOLEAN
@@ -21,7 +21,7 @@ CREATE TABLE IF NOT EXISTS invoices (
     id INTEGER,
     user_id INTEGER,
     contract_name VARCHAR,
-    gold DOUBLE,
+    gold INTEGER,
     expected_usd DOUBLE,
     received_usd DOUBLE,
     timestamp TIMESTAMP
